jts-flask-api-sykr/requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.6.1
EasyProcess==0.2.5
Flask==0.12.3
Flask-Cors==3.0.7
gunicorn==19.6.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysql-connector==2.2.9
numpy==1.14.5
pandas==0.20.3
pickle-mixin==1.0.1
pycparser==2.19
pyOpenSSL==19.0.0
python-dateutil==2.8.0
pytz==2019.1
PyVirtualDisplay==0.2.1
requests==2.21.0
scikit-learn==0.19.1
scipy==0.19.1
selenium==3.141.0
six==1.12.0
soupsieve==1.9.1
urllib3==1.24.3
Werkzeug==0.15.2
WTForms==2.2.1
